tests/wpt/web-platform-tests/css/css-text/parsing/white-space-collapse-invalid copy.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-wrap with invalid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<meta name="assert" content="text-wrap supports only the grammar 'wrap | nowrap | balance | stable | pretty'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("white-space-collapse", "auto");
test_invalid_value("white-space-collapse", "none");
test_invalid_value("white-space-collapse", "collapse preserve");
test_invalid_value("white-space-collapse", "preserve preserve-breaks");
test_invalid_value("white-space-collapse", "delicious collapse");
test_invalid_value("white-space-collapse", "5px");
</script>
</body>
</html>
